skyforge/apps/worker/pyproject.toml
[project]
name = "skyforge-worker"
version = "0.1.0"
description = "Skyforge ODM worker — processes drone imagery via OpenDroneMap"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=2.19.0",
    "google-cloud-pubsub>=2.28.0",
    "httpx>=0.28.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
